# login-federado/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CityPass+ Login Federado — API",
    "description": "Proveedor de identidad (IdP) de CityPass+.\n\n**Autenticación** (`/auth`): login, refresh y logout con LDAP + JWT (RS256).\n**Perfil** (`/me`): acciones autenticadas del usuario sobre su propia cuenta (requiere access token humano).\n**Panel de administración** (`/panel`): ABM de personas y grupos, exclusivo para delegados de módulo (o admin-global con `?module=`).\n**OAuth** (`/oauth/token`): tokens de servicio backend-a-backend (client_credentials).\n**JWKS** (`/.well-known/jwks.json`): clave pública para que otros módulos validen tokens sin llamarnos.\n**Salud** (`/actuator/health`): estado del servicio.\n\nSwagger UI disponible en `/docs`. OpenAPI JSON en `/v3/api-docs`. Puertos: API `8081`, anomaly-detection `8000`, LDAP `389`, Postgres `5432`.",
    "version": "0.1.0-SNAPSHOT",
    "contact": {
      "name": "Grupo 2",
      "email": "grupo2@citypass.local"
    },
    "license": {
      "name": "UADE"
    }
  },
  "servers": [
    { "url": "http://localhost:8081", "description": "Desarrollo (API Java)" },
    { "url": "http://localhost:8000", "description": "Anomaly-detection (FastAPI: GET /health, POST /score)" }
  ],
  "security": [
    { "bearer-jwt": [] }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Login de usuario",
        "description": "Autentica contra LDAP y emite access + refresh token. Los errores de credencial, usuario inexistente, cliente desconocido, módulo cruzado, cuenta deshabilitada, lockout (5 fallos/15 min), bloqueo por anomalías (IA) y anomaly-service caído (fail-closed) devuelven exactamente la misma respuesta 401 (anti-enumeración). El clientId debe corresponder al módulo LDAP del usuario, salvo citypass-admin-web (transversal, emite aud citypass-admin-api).",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access + refresh token emitidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Validación de campos fallida",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas (genérico)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Refresh token",
        "description": "Canjea un refresh token opaco por un par nuevo (rotación RFC 9700). Revalida contra LDAP en cada canje. El reuso de un token ya canjeado revoca toda la cadena.",
        "operationId": "refresh",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Par de tokens rotados",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Validación de campos fallida",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido, revocado o cadena muerta",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Logout",
        "description": "Invalida la sesión (revoca el refresh canjeable). Persiste. El access token vigente sigue vivo hasta 15 minutos: aceptado y esperado. Token desconocido → 204 igual: no se revela si alguna vez existió.",
        "operationId": "logout",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Sesión invalidada (token inexistente devuelve lo mismo)" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Solicitar contraseña temporal",
        "description": "Genera una contraseña aleatoria, la fija en LDAP y la manda por mail. Responde SIEMPRE 204 (usuario existente o no): no se puede enumerar el directorio. Con el mail en mano el usuario entra con esa clave y la cambia en /me/change-password.",
        "operationId": "forgotPassword",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Procesado (el mail llega solo si el usuario existe y tiene email)" },
          "400": {
            "description": "Validación de campos fallida",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por el IdP. Panel requiere: aud=citypass-admin-api, token_use=human, ver=1, claim module y grupo delegados (claim module) O grupo admin-global (módulo vía ?module= en cada request, obligatorio). Sin token → 401; sin claims → 403."
      }
    },
...
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password", "clientId"],
        "properties": {
          "username": { "type": "string", "description": "Nombre de usuario (uid en LDAP)", "example": "jperez" },
          "password": { "type": "string", "description": "Contraseña en texto plano", "example": "changeit123" },
          "clientId": { "type": "string", "description": "client_id de la aplicación (ej: citypass-admin-web, citypass-reclamos-web)", "example": "citypass-admin-web" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string", "description": "JWT RS256, vida 15 min" },
          "refresh_token": { "type": "string", "description": "Token opaco rotativo, vida 8 h" },
          "token_type": { "type": "string", "example": "Bearer" },
          "expires_in": { "type": "integer", "format": "int64", "description": "Segundos hasta expiración del access token", "example": 900 }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string", "description": "Refresh token vigente", "example": "rt_xxxx" }
        }
      },
...
      "AuthErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "status": { "type": "integer", "example": 401 },
          "error": { "type": "string", "example": "Credenciales inválidas" },
          "message": { "type": "string", "example": "Usuario o contraseña inválidos" }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["uid"],
        "properties": {
          "uid": { "type": "string", "description": "UID del usuario (uid en LDAP). Se acepta 'username' como alias legacy.", "example": "jperez" }
        }
      },
...
    }
  }
}